Add support for RV32A in a single-core setting
This adds support for handling instructions specified in the A extension (Zaamo, Zalrsc)
Additions are made to the decoder, pkg, and tracer to decode atomic instructions and
to the LSU. The LSU now contains a combinational AMO ALU to evaluate Zaamo
instructions and an additional state that is required to perform a load and store
required by said instructions. Additionally sequential logic was added to track a
reservation and memory address for the Zalrsc extension.